[Fix] fp32 residual dtype leak in GemmaRMSNorm.forward_native
GemmaRMSNorm.forward_native upcasts x+residual to fp32 for numerical
stability but only casts the output back to orig_dtype, returning the
residual in fp32. This causes dtype mismatches in downstream layers
(e.g. DFlash speculative decoding on ROCm) when the fp32 residual
propagates through the model.

Cast residual back to orig_dtype in the return path, matching the
contract established by RMSNorm.forward_native.

+++ vllm/model_executor/layers/layernorm.py
@@ -167,7 +167,9 @@ class GemmaRMSNorm(CustomOp):
         # ir.ops.rms_norm handles fp32 upcast internally
         out = ir.ops.rms_norm(x, weight, self.variance_epsilon)
         return (
-            out.to(orig_dtype) if residual is None else (out.to(orig_dtype), residual)
+            out.to(orig_dtype)
+            if residual is None
+            else (out.to(orig_dtype), residual.to(orig_dtype))
         )
 
     def forward_cuda(
